Adding ml2_wide funcs

+++ implementations/metalibm/requirements/build.sh
@@ -1,29 +1,25 @@
 #!/bin/bash
-set -x
+
+
 set -e
 
 
-SUCCESS=0
+set -e
 
 SCRIPT_LOCATION="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+
 LOG="${SCRIPT_LOCATION}/log.txt"
 rm -f "${LOG}"
-ACT="${SCRIPT_LOCATION}/activate"
-rm -f "${ACT}"
+echo "START" >> "${LOG}"
 
-pushd () {
-    command pushd "$@" > /dev/null
-}
-
-popd () {
-    command popd > /dev/null
-}
-
+SUCCESS=0
 function finish {
     if [ "$SUCCESS" == 0 ]
     then
-	echo "MetaLibm and requirements failed to build."
+	echo "OpenLibm failed to build."
 	echo "See ${LOG} for details."
+        echo ""
+        tail "${LOG}"
     else
 	echo "Success"
     fi
@@ -31,21 +27,38 @@ function finish {
 trap finish EXIT
 
 
+ACTIVATE_SCRIPT="${SCRIPT_LOCATION}/activate"
+rm -f "${ACTIVATE_SCRIPT}"
+
 # Gappa release
 echo "Installing gappa"
-pushd "$SCRIPT_LOCATION"
-rm -rf gappa-1.3.5 gappa-1.3.5.tar.gz
-wget --no-verbose https://gforge.inria.fr/frs/download.php/file/38044/gappa-1.3.5.tar.gz #&>> "${LOG}"
-tar xf gappa-1.3.5.tar.gz #&>> "${LOG}"
-rm gappa-1.3.5.tar.gz
-pushd gappa-1.3.5
-./configure #&>> "${LOG}"
-./remake #&>> "${LOG}"
+if [ -f "${SCRIPT_LOCATION}/gappa/src/gappa" ]; then
+    echo "  gappa already installed"
+else
+    cd "${SCRIPT_LOCATION}"
 
-export PATH="${PWD}/src:${PATH}"
-echo "export PATH=${PWD}/src:\$PATH" >> "${ACT}"
-popd
-popd
+    echo "  Cleaning build location"
+    rm -rf gappa gappa-1.3.5 gappa-1.3.5.tar.gz
+
+    echo "  Downloading"
+    wget --no-verbose https://gforge.inria.fr/frs/download.php/file/38044/gappa-1.3.5.tar.gz &>> "${LOG}"
+
+    echo "  Untaring"
+    tar xf gappa-1.3.5.tar.gz &>> "${LOG}"
+    mv gappa-1.3.5 gappa
+
+    echo "  Configuring"
+    cd gappa
+    ./configure &>> "${LOG}"
+
+    echo "  Building"
+    ./remake &>> "${LOG}"
+
+    echo "  Done"
+fi
+
+export PATH="${SCRIPT_LOCATION}/gappa/src:${PATH}"
+echo "export PATH=${SCRIPT_LOCATION}/gappa/src:\$PATH" >> "${ACTIVATE_SCRIPT}"
 
 
 
@@ -70,25 +83,37 @@ popd
 
 
 # PythonSollya
-echo "Installing pythonbsollya"
-python3 -m pip install git+https://gitlab.com/metalibm-dev/pythonsollya
+echo "Installing pythonsollya with pip"
+python3 -m pip install git+https://gitlab.com/metalibm-dev/pythonsollya &>> "${LOG}"
 
 
 # asmde
-echo "Installing asmde"
-python3 -m pip install git+https://github.com/nibrunie/asmde
+echo "Installing asmde with pip"
+python3 -m pip install git+https://github.com/nibrunie/asmde  &>> "${LOG}"
 
 
 # MetaLibm
 echo "Installing metalibm"
-pushd "$SCRIPT_LOCATION"
-rm -rf metalibm
-git clone https://github.com/kalray/metalibm.git
-pushd metalibm
-export PYTHONPATH="${PWD}:${PYTHONPATH}"
-echo "export PYTHONPATH=${PWD}:\$PYTHONPATH" >> "${ACT}"
-popd
-popd
+if [ -f "${SCRIPT_LOCATION}/metalibm/metalibm_core/support_lib/ml_support_lib.h" ]; then
+    echo "  metalibm already installed"
+else
+    cd "${SCRIPT_LOCATION}"
+
+    echo "  Cleaning build location"
+    rm -rf metalibm
+
+    echo "  Cloning"
+    git clone https://github.com/kalray/metalibm.git &>> "${LOG}"
+
+    # update time so make will be happy
+    touch "${SCRIPT_LOCATION}/metalibm/metalibm_core/support_lib/ml_support_lib.h"
+    echo "  Done"
+fi
+
+export PYTHONPATH="${SCRIPT_LOCATION}/metalibm:${PYTHONPATH}"
+echo "export PYTHONPATH=${SCRIPT_LOCATION}/metalibm:\$PYTHONPATH" >> "${ACTIVATE_SCRIPT}"
 
 
+
+# Done
 SUCCESS=1
